backend-fastapi/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis for caching and rate limiting
redis==5.0.1
aioredis==2.0.1

# HTTP client for external API calls
httpx==0.25.2
aiohttp==3.9.1

# Data validation and serialization
email-validator==2.1.0
phonenumbers==8.13.27

# Utilities
python-dateutil==2.8.2
pytz==2023.3
ulid-py==1.1.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring and logging
prometheus-client==0.19.0

# Environment and configuration
python-dotenv==1.0.0

# CORS and middleware
starlette==0.27.0

# Additional utilities
typing-extensions==4.8.0

# AWS S3 and storage
boto3==1.34.0
botocore==1.34.0

# Image processing
Pillow==10.1.0
python-magic==0.4.27

# Background task processing
celery==5.3.4
redis==5.0.1

# Additional utilities for media
python-magic-bin==0.4.14
